avoid compiler warning
svn path=/trunk/kdelibs/; revision=77925

+++ kdeui/kdialogbase.cpp
@@ -966,7 +966,7 @@ void KDialogBase::slotNo()
 void KDialogBase::slotCancel()
 {
   emit cancelClicked();
-  done( mMessageBoxMode == true ? Cancel : Rejected );
+  done( mMessageBoxMode == true ? (int)Cancel : (int)Rejected );
 }
